requirements.txt
Django==4.2.11
djangorestframework==3.15.2
django-tenants==3.5.0
psycopg2-binary==2.9.9
anthropic==0.39.0
openai==1.54.0
celery==5.4.0
redis==5.2.0
tiktoken==0.8.0
django-environ==0.11.2
PyPDF2==3.0.1
slack-bolt
django-cors-headers
